requirements.txt
flask==0.10.1
pyserial==2.7
